apps/ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "litestar[standard]>=2.19.0",
    "pydantic>=2.12.5",
    "tortoise-orm[accel,asyncpg]>=0.25.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.5",
]

[tool.uv]
workspace.exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I001",
]
# Unlike Flake8, default to a complexity level of 10.
mccabe.max-complexity = 10

[tool.ruff.format]
line-ending = "lf"
quote-style = "double"
indent-style = "space"
docstring-code-format = true
